+++ src/RcppExports.cpp
@@ -31,8 +31,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // CPP_stats
-Rcpp::NumericMatrix CPP_stats(Rcpp::S4& rast, Rcpp::Nullable<Rcpp::S4> weights, const Rcpp::RawVector& wkb, const Rcpp::StringVector& stats, int max_cells_in_memory);
-RcppExport SEXP _exactextractr_CPP_stats(SEXP rastSEXP, SEXP weightsSEXP, SEXP wkbSEXP, SEXP statsSEXP, SEXP max_cells_in_memorySEXP) {
+Rcpp::NumericMatrix CPP_stats(Rcpp::S4& rast, Rcpp::Nullable<Rcpp::S4> weights, const Rcpp::RawVector& wkb, const Rcpp::StringVector& stats, int max_cells_in_memory, const Rcpp::Nullable<Rcpp::NumericVector>& quantiles);
+RcppExport SEXP _exactextractr_CPP_stats(SEXP rastSEXP, SEXP weightsSEXP, SEXP wkbSEXP, SEXP statsSEXP, SEXP max_cells_in_memorySEXP, SEXP quantilesSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -41,7 +41,21 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< const Rcpp::RawVector& >::type wkb(wkbSEXP);
     Rcpp::traits::input_parameter< const Rcpp::StringVector& >::type stats(statsSEXP);
     Rcpp::traits::input_parameter< int >::type max_cells_in_memory(max_cells_in_memorySEXP);
-    rcpp_result_gen = Rcpp::wrap(CPP_stats(rast, weights, wkb, stats, max_cells_in_memory));
+    Rcpp::traits::input_parameter< const Rcpp::Nullable<Rcpp::NumericVector>& >::type quantiles(quantilesSEXP);
+    rcpp_result_gen = Rcpp::wrap(CPP_stats(rast, weights, wkb, stats, max_cells_in_memory, quantiles));
+    return rcpp_result_gen;
+END_RCPP
+}
+// CPP_resample
+Rcpp::S4 CPP_resample(Rcpp::S4& rast_in, Rcpp::S4& rast_out, const Rcpp::StringVector& stat);
+RcppExport SEXP _exactextractr_CPP_resample(SEXP rast_inSEXP, SEXP rast_outSEXP, SEXP statSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< Rcpp::S4& >::type rast_in(rast_inSEXP);
+    Rcpp::traits::input_parameter< Rcpp::S4& >::type rast_out(rast_outSEXP);
+    Rcpp::traits::input_parameter< const Rcpp::StringVector& >::type stat(statSEXP);
+    rcpp_result_gen = Rcpp::wrap(CPP_resample(rast_in, rast_out, stat));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -49,7 +63,8 @@ END_RCPP
 static const R_CallMethodDef CallEntries[] = {
     {"_exactextractr_CPP_exact_extract", (DL_FUNC) &_exactextractr_CPP_exact_extract, 2},
     {"_exactextractr_CPP_coverage_fraction", (DL_FUNC) &_exactextractr_CPP_coverage_fraction, 3},
-    {"_exactextractr_CPP_stats", (DL_FUNC) &_exactextractr_CPP_stats, 5},
+    {"_exactextractr_CPP_stats", (DL_FUNC) &_exactextractr_CPP_stats, 6},
+    {"_exactextractr_CPP_resample", (DL_FUNC) &_exactextractr_CPP_resample, 3},
     {NULL, NULL, 0}
 };